requirements.txt
Flask==3.0.3
redis==5.0.4
python-dotenv==1.0.1
fakeredis==2.21.3
pytest==8.2.1
requests==2.32.3
